+++ src/main/java/org/organicdesign/fp/collections/UnmodCollection.java
@@ -32,8 +32,6 @@ import java.util.function.Predicate;
 
  UnmodCollection is an unmodifiable version of {@link java.util.Collection}
  which formalizes the return type of Collections.unmodifiableCollection() and Map.values().
-
- {@inheritDoc}
  */
 public interface UnmodCollection<E> extends Collection<E>, UnmodIterable<E> {